feat(task): finish tasks

+++ src/main/java/world/anhgelus/khunegos/player/Task.java
@@ -8,9 +8,12 @@ import java.util.UUID;
 public class Task {
     public final Role role;
     public final UUID linked;
+    public final Prisoner prisoner;
     public boolean win;
+    private boolean running = true;
 
-    public Task(Role role, UUID linked) {
+    public Task(Role role, Prisoner prisoner, UUID linked) {
+        this.prisoner = prisoner;
         this.role = role;
         this.linked = linked;
         win = role == Role.PREY;
@@ -31,39 +34,50 @@ public class Task {
         player.sendMessage(Text.of(sb.toString()));
     }
 
-    public void finishTask(Prisoner prisoner) {
-        prisoner.removeTask(this);
+    public void finishTask() {
         Text message;
         float healthAttribute = 0;
-        switch (role) {
-            case HUNTER:
-                if (win) {
-                    message = Text.of("You killed your prey! You gain one heart!");
-                    healthAttribute = 2.0f;
-                } else {
-                    message = Text.of("You didn't kill your prey! You lose two hearts!");
-                    healthAttribute = -4.0f;
-                }
-                break;
-            case PREY:
-                final var server = prisoner.player().getServer();
-                if (server == null) throw new IllegalStateException("Server is null");
-                final var hunter = server.getPlayerManager().getPlayer(linked);
-                if (hunter == null) throw new IllegalStateException("Player is null (not online?)");
-                final var name = hunter.getDisplayName();
-                assert name != null;
-                if (win) {
-                    message = Text.of("You survived! Your hunter was "+name.getString());
-                } else {
-                    message = Text.of("You were killed by your hunter and you lose five hearts! They were "+
-                            name.getString());
-                    healthAttribute = -10.0f;
-                }
-                break;
-            default:
-                throw new IllegalStateException("Unexpected value: " + role);
+        if (role == Role.HUNTER) {
+            if (win) {
+                message = Text.of("You killed your prey! You gain one heart!");
+                healthAttribute = 2.0f;
+                running = false;
+                prisoner.removeTask(this);
+            } else {
+                message = Text.of("You didn't kill your prey! You lose two hearts!");
+                healthAttribute = -4.0f;
+            }
+        } else if (role == Role.PREY) {
+            // if the player is not linked
+            if (prisoner.uuid == linked) running = false;
+
+            final var server = prisoner.player().getServer();
+            if (server == null) throw new IllegalStateException("Server is null");
+            final var hunter = server.getPlayerManager().getPlayer(linked);
+            if (hunter == null) throw new IllegalStateException("Player is null (not online?)");
+
+            // check if linked hunter's task is running
+            for (final var task : Prisoner.from(hunter).getHunterTasks())
+                if (task.linked == prisoner.uuid && task.running) return;
+
+            final var name = hunter.getDisplayName();
+            assert name != null;
+            if (win) {
+                message = Text.of("You survived! Your hunter was "+name.getString());
+            } else {
+                message = Text.of("You were killed by your hunter and you lose five hearts! They were "+
+                        name.getString());
+                healthAttribute = -10.0f;
+            }
+            running = false;
+        } else {
+            throw new IllegalStateException("Unexpected value: " + role);
         }
         prisoner.player().sendMessage(message);
         prisoner.modifyHealth(healthAttribute);
+    }
+
+    public boolean isRunning() {
+        return running;
     }
 }
